Move struct/enum in espr::parser::subsuper:: to ast::entity::
